fix: token expiration time
