--- src/main/java/api/educai/security/CustomAuthenticationEntryPoint.java
@@ -1,20 +1,0 @@
-package api.educai.security;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import javax.management.BadAttributeValueExpException;
-import javax.naming.InsufficientResourcesException;
-import java.io.IOException;
-
-@Component
-public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException {
-        response.sendError(HttpStatus.BAD_REQUEST.value(), authException.getMessage());
-    }
-}
