+++ src/main/java/api/educai/services/token/RefreshToken.java
@@ -1,9 +1,7 @@
-package api.educai.utils.token;
+package api.educai.services.token;
 
 import api.educai.dto.UserDetailsDTO;
 import api.educai.entities.TokenBlacklist;
-import api.educai.entities.User;
-import api.educai.enums.Role;
 import api.educai.repositories.TokenBlacklistRepository;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.JWTVerifier;
@@ -13,56 +11,52 @@ import com.auth0.jwt.interfaces.DecodedJWT;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
+import java.time.Instant;
+import java.time.LocalDateTime;
+import java.time.ZoneOffset;
 import java.util.Date;
 import java.util.Optional;
 
 @Service
-public class Token implements IToken{
-    @Value("${jwt.token.secret}")
-    private String tokenSecretKey;
+public class RefreshToken implements IToken {
+    @Value("${jwt.refreshToken.secret}")
+    private String refreshSecretKey;
+
     @Autowired
     private TokenBlacklistRepository tokenBlacklistRepository;
 
     @Override
     public String getToken(UserDetailsDTO user) {
         try {
-            long exp = System.currentTimeMillis() + (15 * 60 * 1000); //Expires in 15 minutes
+            Instant exp = LocalDateTime.now().plusDays(15).toInstant(ZoneOffset.of("-03:00")); //Expires in 15 days
 
             return JWT.create()
-                    .withIssuer("educ.ai-api")
                     .withClaim("id", user.getId().toString())
-                    .withClaim("role", user.getRole().toString())
-                    .withClaim("exp", exp)
-                    .sign(Algorithm.HMAC256(tokenSecretKey));
-        } catch (JWTCreationException ex) {
-            throw new RuntimeException("Error while generating access token");
+                    .withExpiresAt(exp)
+                    .sign(Algorithm.HMAC256(refreshSecretKey));
+        } catch(JWTCreationException ex) {
+            throw new RuntimeException("Error while generating refresh token");
         }
     }
 
     @Override
     public ObjectId getUserIdByToken(String token) {
-        JWTVerifier verifier = JWT.require(Algorithm.HMAC256(tokenSecretKey)).build();
+        JWTVerifier verifier = JWT.require(Algorithm.HMAC256(refreshSecretKey)).build();
         DecodedJWT decodedJWT = verifier.verify(token);
 
-        return new ObjectId(decodedJWT.getClaims().get("id").asString());
+        ObjectId userId = new ObjectId(decodedJWT.getClaims().get("id").asString());
+
+        return userId;
     }
 
     @Override
     public Date getTokenExpirationTime(String token) {
-        JWTVerifier verifier = JWT.require(Algorithm.HMAC256(tokenSecretKey)).build();
+        JWTVerifier verifier = JWT.require(Algorithm.HMAC256(refreshSecretKey)).build();
         DecodedJWT decodedJWT = verifier.verify(token);
 
         return decodedJWT.getExpiresAt();
-    }
-
-    public Role getUserRoleByToken(String token) {
-        JWTVerifier verifier = JWT.require(Algorithm.HMAC256(tokenSecretKey)).build();
-        DecodedJWT decodedJWT = verifier.verify(token);
-
-        return Role.valueOf(decodedJWT.getClaims().get("role").asString());
     }
 
     @Override
@@ -73,6 +67,6 @@ public class Token implements IToken{
             return false;
         }
 
-        return tokenBlacklistRepository.existsByIdAndAccessTokenListToken(userId, token);
+        return tokenBlacklistRepository.existsByIdAndRefreshTokenListToken(userId, token);
     }
 }
