--- src/main/java/api/educai/utils/annotations/Student.java
@@ -1,11 +1,0 @@
-package api.educai.utils.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface Student {
-}
